fix: Correción error de casteo (int -> size_t), linea 11

+++ src/mapa.cpp
@@ -1,12 +1,12 @@
 #include "mapa.hpp"
 #include "matriz.hpp"
 bool mapa::es_vecino_valido(coordenada posicion) {
-    int tam_filas=int(map.filas());
-    int tam_col=int(map.columnas());
+    int tam_filas = int(map.filas());
+    int tam_col = int(map.columnas());
 
-    if (posicion.x() < 0 || posicion.x() >= tam_filas || posicion.y() < 0 || posicion.y() >=tam_col) {
+    if (posicion.x() < 0 || posicion.x() >= tam_filas || posicion.y() < 0 || posicion.y() >= tam_col) {
         return false;
     }
 
-    return map.elemento(posicion.x(),posicion.y()) == nullptr;
+    return map.elemento(static_cast<size_t>(posicion.x()), static_cast<size_t>(posicion.y())) == nullptr;
 }
